Rename: 회원가입/로그인관련 변수명 user -> auth로 수정
