+++ src/components/Auth/index.ts
@@ -1,0 +1,6 @@
+import AuthInputBox from './AuthInputBox/AuthInputBox';
+import AuthEmailInputBox from './AuthInputBox/AuthEmailInputBox';
+import AuthPwInputBox from './AuthInputBox/AuthPwInputBox';
+import ErrorMessage from './ErrorMessage';
+
+export { AuthInputBox, AuthEmailInputBox, AuthPwInputBox, ErrorMessage };
