--- src/components/User/index.ts
@@ -1,6 +1,0 @@
-import UserInputBox from './useInputBox/UserInputBox';
-import UserEmailInputBox from './useInputBox/UserEmailInputBox';
-import UserPwInputBox from './useInputBox/UserPwInputBox';
-import ErrorMessage from './ErrorMessage';
-
-export { UserInputBox, UserEmailInputBox, UserPwInputBox, ErrorMessage };
